pyproject.toml
[project]
name = "project-alpha"
version = "0.0.1"
description = "End-to-end data ingestion project with visualization"
authors = [{ name = "Iman", email = "mhdiman17@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.118.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sse>=1.0.0",
    "gtfs-realtime-bindings>=1.0.0",
    "httpx>=0.28.1",
    "kafka-python>=2.2.15",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "repo-to-llm>=0.1.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn>=0.37.0",
]

[project.urls]
Homepage = "https://iman1704.github.io/project-alpha/"
Repository = "https://github.com/iman1704/project-alpha"
Documentation = "https://iman1704.github.io/project-alpha/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    
    "mypy>=0.991",
    
    "ruff>=0.11.5",
    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project_alpha"]

[tool.mypy]
files = ["project_alpha"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.pyrefly]
project-includes = ["project_alpha"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


